add following and follower and posts count to student table and controller add like and comment count to post controller

+++ database/migrations/2019_11_27_220539_create_students_table.php
@@ -14,7 +14,7 @@ class CreateStudentsTable extends Migration
     public function up()
     {
         Schema::create('students', function (Blueprint $table) {
-            $table->bigInteger('phone_number')->unsigned();
+            $table->integer('phone_number')->unsigned();
             $table->string('first_name');
             $table->string('last_name')->nullable();
             $table->string('username');
